reordering files to make possible growing app
