+++ src/routes/middlewares/auth.js
@@ -1,0 +1,14 @@
+const isAuthenticated = (req, res) => {
+  // verify if auth
+};
+
+const signUp = (req, res) => {
+  //signUp
+  //create user in db
+};
+
+const signIn = (req, res) => {
+  //signIn
+};
+
+module.export = { isAuthenticated, signUp, signIn };
